+++ src/loaders/index.js
@@ -2,6 +2,7 @@ const expressLoader = require('./express')
 const databaseLoader = require('./database')
 const routeLoader = require('./route')
 const uploadLoader = require('./uploads')
+const passportLoader =require('./passport')
 // const jobLoader = require('./jobs')
 
 async function init(app) {
@@ -13,6 +14,10 @@ async function init(app) {
       // 데이터베이스 초기화
       await databaseLoader()
       console.log('✅ 데이터베이스 초기화 완료')
+
+      await passportLoader(app)
+      console.log('✅ 데이터베이스 초기화 완료')
+
 
       // 업로드 디렉토리 초기화
       await uploadLoader()